requirements.txt
numpy==1.21.6
pandas==1.3.5
cloudpickle==2.2.0
unidecode==1.3.6
scikit-learn==1.0.2
flask==2.2.2
tensorflow-cpu==2.11.0
requests==2.28.1
flask_cors==3.0.10
gunicorn==20.1.0